requirements.txt
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2
joblib==1.5.2
streamlit==1.50.0
plotly==6.3.0
scipy==1.16.2
tensorflow==2.20.0
pillow==11.3.0
tensorflow-hub==0.16.1